Splitted http_transporter from transporter.go
